backend/requirements.txt
Flask
Flask-CORS
xhtml2pdf
mysql-connector-python
openpyxl
reportlab
Werkzeug
python-dotenv